Add loading module prefab from UI_LoadingScreen
