+++ nekoyume/Assets/AddressableAssets/UI/Module/LoadingModule.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 9387536cf1f3c094f8af104689eb03a6
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
